+++ src/navigation/AppNavigator.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 
 import MainTabNavigator from './MainTabNavigator';
-import LoadingScreen from '../screens/LoadingScreen';
-import LoginScreen from '../screens/LoginScreen';
-import DashboardScreen from '../screens/DashboardScreen';
+import LoadingScreen from '../../screens/LoadingScreen';
+import LoginScreen from '../../screens/LoginScreen';
+import DashboardScreen from '../../screens/DashboardScreen';
 
 
 export default createAppContainer(
